refactor project structure
New module structure and some deleted components
