+++ src/app/communityDashboard/page.tsx
@@ -1,4 +1,4 @@
-import Dashboard from "@/components/dasboard";
+import Dashboard from "@/app/shared/components/dasboard";
 
 export default function DashboardPage() {
   return <Dashboard />;
