+++ src/app/shared/components/CommunityCard.tsx
