+++ src/app/shared/components/avatar.tsx
