+++ src/app/shared/components/card.tsx
