+++ src/app/shared/components/community-banner.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { Button } from "@/components/ui/button"
+import { Button } from "@/app/shared/components/button"
 import { Globe2, Users } from 'lucide-react'
 
 interface CommunityBannerProps {
