+++ src/app/shared/components/highlighted-vote.tsx
@@ -1,5 +1,5 @@
-import { Badge } from "./ui/badge"
-import { Card, CardContent, CardHeader } from "./ui/card"
+import { Badge } from "./badge"
+import { Card, CardContent, CardHeader } from "./card"
 
 interface HighlightedVoteProps {
   question: string
