+++ src/app/shared/components/highlighted-votes-carousel.tsx
