+++ src/app/shared/components/post-form.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { Button } from "@/components/ui/button";
+import { Button } from "@/app/shared/components/button";
 import {
   ChevronDown,
   ChevronUp,
